refactor(nav): extract CompassController from ExperimentalControls
Move the compass + plan-view controller into compassController.js: the top-down
plan-view framing tween, the compass body-click / rotate-arrow actions, the
shared yaw-about-pivot tween primitive, and the at-most-one queued-action drain.
These motions keep their own TickAnimator subscription + ownership flags
(planViewActive / compassAnimating feed the orchestrator's _isInactive via
ctx.compass) — they do NOT enter the committed-motion runner.

The compass math helpers (needleScreenAngle / signToYaw / normalizeDeg) move here
with it; needleScreenAngle is re-exported from ExperimentalControls.js so the
Compass widget's frozen import path (via index.js) is unchanged. The compass /
plan-view camera writes now commit through the write funnel (invalidate zoom-undo
+ dispatch), completing the funnel migration for these sites. ctx gains bounds,
disabledByOrtho, compass; the orchestrator keeps handlePlanViewRequest /
handleCompassBodyClick / handleCompassRotate as public delegators.

Behaviour-preserving: 302/302 characterization tests green.
